Fix build and remove Emacs support
The build was using rsync to move files to a directory outside the repo.
Now, it just uses GOPATH to direct the go tools to the files inside the
repo. The old setup probably worked nowhere but on the old machine I was
using at the time.

I've added a topological sort package to the repo as it was a hidden
dependency that should have been in the repo the whole time but just
wasn't.

I removed Emacs support because I don't use Emacs any more. Better to
remove it than let the bitrot get to it.

There are a few other small changes in this commit, like an update to
the README and the .gitignore.