Add an anonymous 'let' binding form
This new form is useful for introducing a fresh local scope. Example:

(let _
   (let x 6)
   (let y 7)
   (* x y))

In traditional Scheme, the above would be written:

(let ((x 6)
      (y 7))
  (* x y))

The new version uses more 'ink' but I think it's a good trade-off.

One neat feature of this new syntax is that if you decide you want to name the
value, you just have to replace the '_' with a name.

I have more variations on this theme in mind for future commits.