Change syntax again
I started to dislike the overloading of 'func' and 'subr' to be both expression
forms and binding forms so I reverted back to a very simple pattern of having
only 'let' for binding locals and using 'fun' and 'sub' only as expression
forms--now shortened to 3 letters instead of 4.

Downsides include an extra line and an extra indent level in most
function/subroutine definitions.

However, the orthogonality and consistency are appealing.

Aside: I found that the '_' characters stood out too much and wanted something
less distracting. I tried '@', which did seem better but I still disliked the
overloading.

I could introduce binding keywords different from 'let', 'fun', and 'sub'. That
would allow me to keep the concise variant for function/subroutine binders. I
haven't done that because no keywords came to mind and it makes sense,
conceptually, for there just to be 'let'.